Feat(#9): Refactored Instance module to new module structure
