+++ src/commands/instance/common/mod.rs
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2025 Tim Anhalt (BitTim)
+ *
+ * Project:    Minepatch
+ * License:    GPLv3
+ *
+ * File:       mod.rs
+ * Author:     Tim Anhalt (BitTim)
+ * Modified:   13.01.25, 21:32
+ */
+pub mod confirm;
+pub mod registry;
