+++ src/main.rs
@@ -6,15 +6,14 @@
  *
  * File:       main.rs
  * Author:     Tim Anhalt (BitTim)
- * Modified:   13.01.25, 20:29
+ * Modified:   13.01.25, 21:40
  */
 use clap::Parser;
-use minepatch::commands::instance::instance_cli::InstanceCommands;
-use minepatch::commands::instance::instance_main;
+use minepatch::commands::instance::cli::InstanceCommands;
 use minepatch::commands::pack::cli::PackCommands;
 use minepatch::commands::vault::vault_cli::VaultCommands;
 use minepatch::commands::vault::vault_main;
-use minepatch::commands::{pack, update, Cli, Commands};
+use minepatch::commands::{instance, pack, update, Cli, Commands};
 use minepatch::common::error;
 
 fn match_command(command: &Commands) -> error::Result<()> {
@@ -23,10 +22,14 @@ fn match_command(command: &Commands) -> error::Result<()> {
         Commands::Instance {
             instance_commands: instance_command,
         } => match instance_command {
-            InstanceCommands::List => instance_main::list()?,
-            InstanceCommands::Link { path, name } => instance_main::link(path, name)?,
-            InstanceCommands::Rename { name, new_name } => instance_main::rename(name, new_name)?,
-            InstanceCommands::Unlink { name, all, yes } => instance_main::unlink(name, all, yes)?,
+            InstanceCommands::List => instance::func::list::list()?,
+            InstanceCommands::Link { path, name } => instance::func::link::link(path, name)?,
+            InstanceCommands::Rename { name, new_name } => {
+                instance::func::rename::rename(name, new_name)?
+            }
+            InstanceCommands::Unlink { name, all, yes } => {
+                instance::func::unlink::unlink(name, all, yes)?
+            }
         },
         Commands::Vault {
             vault_commands: vault_command,
